+++ src/components/RichTextEditor/index.scss
